requirements.txt
open-clip-torch==2.26.1
transformers==4.44.2
pillow==10.4.0
pandas==2.2.2
faiss-cpu==1.8.0
accelerate==0.34.2